Remove the dead AFT fitter and two more verbatim duplicates
A second duplication sweep, comparing normalised function bodies rather
than names across the package's 1,483 non-trivial functions.

The legacy accelerated_failure_time/accelerated_failure_time.py was
dead code -- the package __init__ re-exports from aft_fitter, nothing
imported the old AcceleratedFailureTimeFitter, and no test touched it.
It carried the only live copy of _parameter_initialiser_dist; the
verbatim copies on the PH fitter and the parameter-substitution fitter
had no callers at all. All three are deleted.

The Bernoulli / FixedEventProbability split had copied fit, from_params,
entropy and random wholesale -- the largest exact duplicate in the
package. They now share SingleProbabilityMixin (ratcheted from birth);
everything distributional stays on the classes. This also fixes
FixedEventProbability.from_params's docstring, which still said 'Create
a Bernoulli model'.

The four-branch support-respecting Wald transform in param_cb was
verbatim between the recurrent inference mixin and the parametric
regression model; it is now utils.linalg.wald_bound_on_support and both
callers delegate.

47 fingerprints are bit-identical against develop, with the baseline
run verified to import the pre-change code. sf_tvc and _prepare_Z were
investigated and deliberately left: thin wrappers over already-shared
machinery, with per-family docstrings worth keeping.

Incidentally, a package-wide isort run fixed the import ordering in ten
files with pre-existing violations (one- to three-line blank-line and
grouping changes; the open isort housekeeping item shrinks accordingly).
