+++ surpyval/univariate/parametric/distributions/fixed_event_probability.py
@@ -16,9 +16,6 @@ constant, so there is no density, no invertible quantile, and no time to
 average.
 """
 
-import numpy.typing as npt
-from scipy.stats import uniform
-
 from surpyval import np
 from surpyval.univariate.parametric.discrete_fitter import (
     DiscreteParametricFitter,
@@ -26,13 +23,12 @@ from surpyval.univariate.parametric.discrete_fitter import (
 from surpyval.univariate.parametric.parametric_fitter import (
     Boxable,
     Numeric,
-    reject_structural_params,
 )
 
-from ..parametric import Parametric
+from ._single_probability import SingleProbabilityMixin
 
 
-class FixedEventProbability_(DiscreteParametricFitter):
+class FixedEventProbability_(SingleProbabilityMixin, DiscreteParametricFitter):
     def __init__(self, name: str) -> None:
         super().__init__(
             name=name,
@@ -183,93 +179,6 @@ class FixedEventProbability_(DiscreteParametricFitter):
         0.5
         """
         return p
-
-    def entropy(self, p: Boxable) -> Boxable:
-        return -(1 - p) * np.log1p(-p) - p * np.log(p)
-
-    def random(self, size: int | tuple[int, ...], p: Boxable) -> npt.NDArray:
-        r"""
-
-        Draws random samples from the distribution in shape `size`
-
-        Parameters
-        ----------
-
-        size : integer or tuple of positive integers
-            Shape or size of the random draw
-        p : float
-            The probability of failure of the thing
-
-        Returns
-        -------
-
-        random : scalar or numpy array
-            Random values drawn from the distribution in shape `size`
-
-        """
-        U = uniform.rvs(size=size)
-        return (U <= p).astype(int)
-
-    def fit(
-        self, x: npt.ArrayLike, n: npt.NDArray | None = None
-    ) -> Parametric:
-        x_arr = np.atleast_1d(x)
-        # Each observation must be a 0 or a 1 — elementwise, for any length
-        # (the previous check broadcast x against the literal [0, 1], so any
-        # input of length != 2 crashed and [1, 1] was rejected, #257).
-        if not np.isin(x_arr, (0, 1)).all():
-            raise ValueError("'x' must be either 0 or 1")
-        n_arr = np.ones_like(x_arr) if n is None else np.atleast_1d(n)
-        if n_arr.shape[0] != x_arr.shape[0]:
-            raise ValueError("'n' must be the same length as 'x'")
-
-        model = Parametric(self, "MLE", None, False, False, False)
-        p = (x_arr * n_arr).sum() / n_arr.sum()
-        model.params = np.array([p])
-        return model
-
-    # Narrower than ParametricFitter.from_params, which takes
-    # (params, gamma, p, f0). Unlike `fit`, this one is not resolved
-    # by the OptimisedFitMixin split: every distribution has a
-    # from_params. It is a parameter *rename* -- the base's `params`
-    # became `p` -- so positional calls work and keyword calls
-    # raise. Worse here: the base's `p` means the
-    # limited-failure proportion, so the same keyword means two
-    # unrelated things across sibling classes. Fixing it means renaming
-    # back, with a deprecation alias, and is tracked separately.
-    def from_params(
-        self,
-        params: npt.ArrayLike,
-        gamma: Boxable | None = None,
-        p: Boxable | None = None,
-        f0: Boxable | None = None,
-    ) -> Parametric:
-        """Create a Bernoulli model from its event probability.
-
-        Parameters
-        ----------
-        params : scalar
-            The event probability, between 0 and 1.
-        gamma, p, f0 : None
-            Accepted so the signature matches
-            :meth:`ParametricFitter.from_params`, and rejected: a
-            Bernoulli has no offset, limited failure population or zero
-            inflation. Note that the base's ``p`` is the *never-fails*
-            proportion, not this distribution's parameter -- which is why
-            the parameter is ``params`` and not ``p``.
-        """
-        reject_structural_params(self.name, gamma, p, f0)
-        prob = float(np.squeeze(np.asarray(params)))
-
-        if prob > 1:
-            raise ValueError("'params' must be less than 1")
-
-        if prob < 0:
-            raise ValueError("'params' must be greater than 0")
-
-        model = Parametric(self, "given parameters", None, False, False, False)
-        model.params = np.atleast_1d(prob)
-        return model
 
 
 FixedEventProbability: FixedEventProbability_ = FixedEventProbability_(
