+++ surpyval/univariate/regression/accelerated_failure_time/aft_fitter.py
@@ -1,8 +1,6 @@
 import autograd.numpy as np
 import numpy.typing as npt
 
-
-from .._likelihood import regression_neg_ll
 from .._fit_skeleton import (
     HazardIdentitiesMixin,
     LogLinearPhi,
@@ -10,6 +8,7 @@ from .._fit_skeleton import (
     optimise_nm_tnc,
     prepare_regression_fit,
 )
+from .._likelihood import regression_neg_ll
 from ..parametric_regression_model import ParametricRegressionModel
 from ..regression_data import DataFrameRegressionMixin
 from .aft_tvc_fit import AFTTVCFitMixin
